packages/backend/requirements.txt
aiofiles>=24.1.0
arize-phoenix-otel>=0.12.1
aws-lambda-powertools>=3.16.0
aws-xray-sdk>=2.14.0
colorama>=0.4.6
fastapi>=0.116.0
ipykernel>=6.29.5
langchain-aws>=0.2.27
langchain-core>=0.3.68
langchain-mcp-adapters>=0.1.9
langgraph>=0.5.2
langgraph-checkpoint-sqlite>=2.0.10
openinference-instrumentation-langchain>=0.1.46
pillow>=11.3.0
pip-licenses>=5.0.0
python-dotenv>=1.1.1
pyyaml>=6.0.2
uvicorn>=0.35.0
